move LegacyTableRow to a more intuitive location
